Fix Cache component regressions for PHPUnit 11 and PHP 8.4
Restore pack/unpack and marshaller wiring dropped in security hardening, fix TagAwareMarshaller null handling, and align adapter tests with typed skippedTests.

+++ src/Symfony/Component/Cache/Adapter/ArrayAdapter.php
@@ -397,6 +397,6 @@ class ArrayAdapter implements AdapterInterface, CacheInterface, NamespacedPoolIn
 
     private function getCurrentTime(): float
     {
-        return $this->clock?->now()->format('U.u') ?? microtime(true);
+        return (float) ($this->clock?->now()->format('U.u') ?? microtime(true));
     }
 }
